requirements.txt
django
gunicorn
django-heroku
djangorestframework
requests
django-cors-headers
networkx
whitenoise
numpy
gensim
interruptingcow 
